requirements.txt
Flask==3.0.0
Pillow==10.1.0
gunicorn==21.2.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
scikit-image==0.22.0